Fix block height in dashboard, reverse title bar to the OS one.
Fixes:
Fix displaying block height on dashboard..

Updates:
Changed the status bar of the wallet from a custom one to the the OS one, preventing on some OS'es the random movement of the wallet when dragged.
Changed the scale option up to 1.9 for better visualization on small displays.
Changed data fetch from nebula homepage to nebula API.

Removes:
Remove the possibility to drag the window from the upper part of the wallet after login.