Fixes and Addings
Fix staking approval message, showing to approve to stake the balance and not the desired tokens count, actual staking tokens not affected.
Fix wallet unexpectedly closing if you click "STAKE" button when no accounts are present in the wallet.

Add dialog message if create dividends was successfully executed.
Add the option to connect the wallet to a custom node IP ( for example: your node IP ), if let empty will connect to the hard written node IP's.